+++ R/data.R
@@ -4,15 +4,18 @@
 #' package. Structured as a list of lists with the first level being each
 #' package and the second level being the palettes available in that package.
 #'
-#' Contains palettes from the following packages: awtools, dichromat,
-#' dutchmasters, ggsci, ggpomological, ggthemes, ghibli, grDevices, jcolors,
-#' LaCroixColoR, NineteenEightyR, nord, ochRe, pals, Polychrome, quickpalette,
-#' rcartocolor, RColorBrewer, Redmonder, RSkittleBrewer, tidyquant, wesanderson
-#' and yarrr.
+#' Contains palettes from the following packages: awtools, basetheme,
+#' calecopal, colorblindr, colRoz, dichromat, dutchmasters, DresdenColor,
+#' fishualize, ggsci, ggpomological, ggthemes, ggthemr, ghibli, grDevices,
+#' IslamicArt, jcolors, LaCroixColoR, lisa, nationalparkcolors,
+#' NineteenEightyR, nord, ochRe, pals, palettesForR_pals, Polychrome,
+#' MapPalettes, miscpalettes, PNWColors, rcartocolor, RColorBrewer, Redmonder,
+#' RSkittleBrewer, tidyquant, trekcolors, tvthemes, unikn, vapeplot, vapoRwave,
+#' werpals, wesanderson and yarrr.
 #'
 #' data.frame of palette names can be found here \code{\link{palettes_d_names}}.
 #'
-#' @format A list of 24 lists.
+#' @format A list of 43 lists.
 "palettes_d"
 
 #' Names of all fixed discrete palettes
@@ -20,15 +23,18 @@
 #' A data.frame of the names and other characteristics of the discrete palettes
 #' of fixed lengths included in this package.
 #'
-#' Contains palettes names from the following packages: awtools, dichromat,
-#' dutchmasters, ggsci, ggpomological, ggthemes, ghibli, grDevices, jcolors,
-#' LaCroixColoR, NineteenEightyR, nord, ochRe, pals, Polychrome, quickpalette,
-#' rcartocolor, RColorBrewer, Redmonder, RSkittleBrewer, tidyquant, wesanderson
-#'  and yarrr.
+#' Contains palettes names from the following packages: awtools, basetheme,
+#' calecopal, colorblindr, colRoz, dichromat, dutchmasters, DresdenColor,
+#' fishualize, ggsci, ggpomological, ggthemes, ggthemr, ghibli, grDevices,
+#' IslamicArt, jcolors, LaCroixColoR, lisa, nationalparkcolors,
+#' NineteenEightyR, nord, ochRe, pals, palettesForR_pals, Polychrome,
+#' MapPalettes, miscpalettes, PNWColors, rcartocolor, RColorBrewer, Redmonder,
+#' RSkittleBrewer, tidyquant, trekcolors, tvthemes, unikn, vapeplot, vapoRwave,
+#' werpals, wesanderson and yarrr.
 #'
 #' Full list of palette can be found here \code{\link{palettes_d}}.
 #'
-#' @format A data.frame of 855 observations with 4 variables:
+#' @format A data.frame of 1419 observations with 4 variables:
 #' \describe{
 #'   \item{package}{character, name of package}
 #'   \item{palette}{character, name of palette}
@@ -76,13 +82,25 @@
 #' palettes included in this package.
 #'
 #' Contains palettes names from the following packages: gameofthrones,
-#' ggthemes, grDevices, harrypotter, jcolors, oompaBase, palr, pals,
+#' ggthemes, grDevices, harrypotter, oompaBase, palr, pals,
 #' scico and viridis.
 #'
-#' @format A data.frame of 205 observations with 3 variables:
+#' @format A data.frame of 315 observations with 3 variables:
 #' \describe{
 #'   \item{package}{character, name of package}
 #'   \item{palette}{character, name of palette}
 #'   \item{type}{character, type of palette}
 #' }
 "palettes_c_names"
+
+#' Names and version information for all packages included
+#'
+#' @format A data.frame with 3 variables:
+#' \describe{
+#'   \item{Names}{character, name of package}
+#'   \item{Github}{character, name of Github repository}
+#'   \item{github_ver}{character, version number on Github}
+#'   \item{CRAN}{logical, is package available on CRAN}
+#'   \item{CRAN_ver}{character, version number on CRAN}
+#' }
+"paletteer_packages"